+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-## 1.0.0
+v1.0.1
+- Modify SSH authentication protocol to use Keyboard Authentication
+- Update Rencii.SshNet version
+
+v1.0.0
 - Initial Version